Moved the entities to a written package because they are complex enough to be tested
