+++ src/main/java/de/uni_stuttgart/it_rex/course/domain/written/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * written JPA domain objects.
+ */
+package de.uni_stuttgart.it_rex.course.domain.written;
