--- src/main/java/de/uni_stuttgart/it_rex/course/domain/written_entities/package-info.java
@@ -1,4 +1,0 @@
-/**
- * written JPA domain objects.
- */
-package de.uni_stuttgart.it_rex.course.domain.written_entities;
