+++ src/main/java/de/uni_stuttgart/it_rex/course/repository/written/LecturerRepository.java
@@ -1,6 +1,6 @@
 package de.uni_stuttgart.it_rex.course.repository.written;
 
-import de.uni_stuttgart.it_rex.course.domain.written_entities.Lecturer;
+import de.uni_stuttgart.it_rex.course.domain.written.Lecturer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
