+++ src/test/java/de/uni_stuttgart/it_rex/course/service/mapper/written/CourseMapperIT.java
@@ -2,7 +2,7 @@ package de.uni_stuttgart.it_rex.course.service.mapper.written;
 
 import de.uni_stuttgart.it_rex.course.CourseServiceApp;
 import de.uni_stuttgart.it_rex.course.config.TestSecurityConfiguration;
-import de.uni_stuttgart.it_rex.course.domain.written_entities.Course;
+import de.uni_stuttgart.it_rex.course.domain.written.Course;
 import de.uni_stuttgart.it_rex.course.service.dto.written_dtos.CourseDTO;
 import de.uni_stuttgart.it_rex.course.utils.written.CourseUtil;
 import org.junit.jupiter.api.Test;
