+++ src/test/java/de/uni_stuttgart/it_rex/course/service/mapper/written/TimePeriodMapperIT.java
@@ -2,7 +2,7 @@ package de.uni_stuttgart.it_rex.course.service.mapper.written;
 
 import de.uni_stuttgart.it_rex.course.CourseServiceApp;
 import de.uni_stuttgart.it_rex.course.config.TestSecurityConfiguration;
-import de.uni_stuttgart.it_rex.course.domain.written_entities.TimePeriod;
+import de.uni_stuttgart.it_rex.course.domain.written.TimePeriod;
 import de.uni_stuttgart.it_rex.course.service.dto.written_dtos.TimePeriodDTO;
 import de.uni_stuttgart.it_rex.course.utils.written.TimePeriodUtil;
 import org.junit.jupiter.api.Test;
