+++ src/test/java/de/uni_stuttgart/it_rex/course/utils/written/ChapterUtil.java
@@ -1,6 +1,6 @@
 package de.uni_stuttgart.it_rex.course.utils.written;
 
-import de.uni_stuttgart.it_rex.course.domain.written_entities.Chapter;
+import de.uni_stuttgart.it_rex.course.domain.written.Chapter;
 import de.uni_stuttgart.it_rex.course.service.dto.written_dtos.ChapterDTO;
 
 import java.util.List;
